Update README.md
##Summary

I am excited to share the first Python project with the group. This was a Python web scraping program, which required importing the Python Beautiful Soup library. The goal was to start off by targeting different zip codes and return a list of the donor amount and their name for Arlington, Virginia.

https://www.vpap.org/zipcodes/zcta/22209/donors/?zip=22209&start_year=2018&end_year=2019&recip_type=all

I would like to express my continued gratitude to Eric, who wrote the code, and explained the request and CSS to capture metadata and xml. This was cumbersome because the first attempt was done using the Python library Scrapy. Beautiful Soup turned out to be more conducive for this program but also laborious.

My friends and I will use the information to target neighborhoods for campaign events promoting Parisa Dehghani-Tafti (D) in Arlington, Virginia. Parisa wants to make court data statistics more available to the public. For example, this will allow analysts to visualize the disparities in arrests by zip code.

https://parisaforjustice.com (feel free to contact me if you want to get involved!)